+++ lib/common/global_variables.dart
@@ -1,12 +1,10 @@
-part of draggable_grid_view;
+part of '../flutter_draggable_gridview.dart';
 
-// var dragStarted = false;
-// var dragEnded = true;
-// late List<DraggableGridItem> orgList;
-// // late List<DraggableGridItem> _list;
-// DraggableGridItem? draggedGridItem;
-// //
-// // /// [isOnlyLongPress] is Accepts 'true' and 'false'
-// // /// If, it is true then only draggable works with long press.
-// // /// and if it is false then it works with simple press.
-// bool isOnlyLongPress = true;
+bool _dragStarted = false;
+bool _dragEnded = true;
+DraggableGridItem? _draggedGridItem;
+
+// /// [isOnlyLongPress] is Accepts 'true' and 'false'
+// /// If, it is true then only draggable works with long press.
+// /// and if it is false then it works with simple press.
+bool _isOnlyLongPress = true;