+++ lib/flutter_draggable_gridview.dart
@@ -2,27 +2,28 @@ library draggable_grid_view;
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-
-import 'constants/colors.dart';
+import 'package:flutter_draggable_gridview/constants/colors.dart';
 
 part 'common/global_variables.dart';
+
 part 'models/draggable_grid_item.dart';
+
 part 'widgets/drag_target_grid.dart';
+
 part 'widgets/empty_item.dart';
+
 part 'widgets/long_press_draggable_grid.dart';
+
 part 'widgets/placeholder_widget.dart';
+
 part 'widgets/press_draggable_grid.dart';
 
-typedef DragCompletion = void Function(
-    List<DraggableGridItem> list, int beforeIndex, int afterIndex);
+typedef DragCompletion = void Function(List<DraggableGridItem> list, int beforeIndex, int afterIndex);
 typedef DragFeedback = Widget Function(List<DraggableGridItem> list, int index);
-typedef DragChildWhenDragging = Widget Function(
-    List<DraggableGridItem> list, int index);
-typedef DragPlaceHolder = PlaceHolderWidget Function(
-    List<DraggableGridItem> list, int index);
+typedef DragChildWhenDragging = Widget Function(List<DraggableGridItem> list, int index);
+typedef DragPlaceHolder = PlaceHolderWidget Function(List<DraggableGridItem> list, int index);
 
 class DraggableGridViewBuilder extends StatefulWidget {
-  ///TestComment
   /// [children] will show the widgets in Gridview.builder.
   final List<DraggableGridItem> children;
 
@@ -62,7 +63,7 @@ class DraggableGridViewBuilder extends StatefulWidget {
   final bool sliverMode;
 
   const DraggableGridViewBuilder({
-    Key? key,
+    super.key,
     required this.gridDelegate,
     required this.children,
     required this.dragCompletion,
@@ -87,25 +88,26 @@ class DraggableGridViewBuilder extends StatefulWidget {
     this.restorationId,
     this.clipBehavior = Clip.hardEdge,
     this.sliverMode = false,
-  }) : super(key: key);
+  });
 
   @override
-  DraggableGridViewBuilderState createState() =>
-      DraggableGridViewBuilderState();
+  DraggableGridViewBuilderState createState() => DraggableGridViewBuilderState();
 }
 
 class DraggableGridViewBuilderState extends State<DraggableGridViewBuilder> {
-  bool _isOnlyLongPress = true;
-  late List<DraggableGridItem> _list;
-  DraggableGridItem? _draggedGridItem;
   late List<DraggableGridItem> _orgList;
+  late List<DraggableGridItem> _list;
+
+  /// [isOnlyLongPress] is Accepts 'true' and 'false'
+  /// If, it is true then only draggable works with long press.
+  /// and if it is false then it works with simple press.
 
   @override
   void initState() {
     super.initState();
     assert(widget.children.isNotEmpty, 'Children must not be empty.');
 
-    /// [list] will update when the widget is beign dragged.
+    /// [list] will update when the widget is begin dragged.
     _list = [...widget.children];
 
     /// [orgList] will set when the drag completes.
@@ -127,30 +129,33 @@ class DraggableGridViewBuilderState extends State<DraggableGridViewBuilder> {
   Widget build(BuildContext context) {
     if (widget.sliverMode) {
       return SliverGrid.builder(
+          key: ValueKey(_orgList),
           itemCount: _list.length,
           gridDelegate: widget.gridDelegate,
           itemBuilder: ((context, index) {
             return (!_list[index].isDraggable)
                 ? _list[index].child
                 : DragTargetGrid(
+                    isOnlyLongPress: _isOnlyLongPress,
+                    list: _list,
+                    orgList: _orgList,
                     index: index,
                     onChangeCallback: () => setState(() {}),
                     feedback: widget.dragFeedback?.call(_list, index),
-                    childWhenDragging:
-                        widget.dragChildWhenDragging?.call(_orgList, index),
+                    childWhenDragging: widget.dragChildWhenDragging?.call(_orgList, index),
                     placeHolder: widget.dragPlaceHolder?.call(_orgList, index),
                     dragCompletion: widget.dragCompletion,
-                    list: const [],
-                    draggedGridItem: null,
-                    onDragGridItem: (item) {},
-                    orgList: const [],
-                    onOrgListUpdate: (data) {},
-                    onListUpdate: (data) {},
-                    isOnlyLongPress: false,
+                    onListUpdate: (List<DraggableGridItem> value) {
+                      _list = value;
+                    },
+                    onOrgListUpdate: (List<DraggableGridItem> value) {
+                      _orgList = value;
+                    },
                   );
           }));
     }
     return GridView.builder(
+      key: ValueKey(_orgList),
       scrollDirection: widget.scrollDirection,
       reverse: widget.reverse,
       controller: widget.controller,
@@ -171,26 +176,20 @@ class DraggableGridViewBuilderState extends State<DraggableGridViewBuilder> {
         return (!_list[index].isDraggable)
             ? _list[index].child
             : DragTargetGrid(
-                index: index,
-                onChangeCallback: () => setState(() {}),
-                feedback: widget.dragFeedback?.call(_list, index),
-                childWhenDragging:
-                    widget.dragChildWhenDragging?.call(_orgList, index),
-                placeHolder: widget.dragPlaceHolder?.call(_orgList, index),
-                dragCompletion: widget.dragCompletion,
                 isOnlyLongPress: _isOnlyLongPress,
                 list: _list,
                 orgList: _orgList,
-                draggedGridItem: _draggedGridItem,
-                onDragGridItem: (DraggableGridItem? data) {
-                  _draggedGridItem = data;
-                  setState(() {});
+                index: index,
+                onChangeCallback: () => setState(() {}),
+                feedback: widget.dragFeedback?.call(_list, index),
+                childWhenDragging: widget.dragChildWhenDragging?.call(_orgList, index),
+                placeHolder: widget.dragPlaceHolder?.call(_orgList, index),
+                dragCompletion: widget.dragCompletion,
+                onListUpdate: (List<DraggableGridItem> value) {
+                  _list = value;
                 },
-                onOrgListUpdate: (List<DraggableGridItem> data) {
-                  _orgList = data;
-                },
-                onListUpdate: (List<DraggableGridItem> data) {
-                  _list = data;
+                onOrgListUpdate: (List<DraggableGridItem> value) {
+                  _orgList = value;
                 },
               );
       },