+++ lib/widgets/empty_item.dart
@@ -1,8 +1,8 @@
-part of draggable_grid_view;
+part of '../flutter_draggable_gridview.dart';
 
 /// [EmptyItem] will use to show at drag target, when the widget is being dragged.
 class EmptyItem extends StatelessWidget {
-  const EmptyItem({Key? key}) : super(key: key);
+  const EmptyItem({super.key});
 
   @override
   Widget build(BuildContext context) {