+++ lib/widgets/long_press_draggable_grid.dart
@@ -1,6 +1,5 @@
-part of draggable_grid_view;
+part of '../flutter_draggable_gridview.dart';
 
-//ignore: must_be_immutable
 class LongPressDraggableGridView extends StatelessWidget {
   /// [index] is use to get item from the list.
   final int index;
@@ -10,53 +9,37 @@ class LongPressDraggableGridView extends StatelessWidget {
 
   /// [DragChildWhenDragging] this to display the widget at dragged widget place when the widget is being dragged.
   final Widget? childWhenDragging;
-  List<DraggableGridItem> list;
 
   final VoidCallback onDragCancelled;
-  bool dragStarted;
-  bool dragEnded;
 
-  Function(bool) onDragStarted;
-  Function(bool) onDragEnded;
+  final List<DraggableGridItem> list;
 
-  DraggableGridItem? draggedGridItem;
-
-  LongPressDraggableGridView({
-    required this.dragEnded,
-    required this.dragStarted,
-    required this.onDragEnded,
-    required this.onDragStarted,
-    required this.draggedGridItem,
-    required this.index,
-    required this.list,
-    required this.onDragCancelled,
-    this.feedback,
-    this.childWhenDragging,
-    Key? key,
-  }) : super(key: key);
+  const LongPressDraggableGridView(
+      {required this.index,
+      required this.onDragCancelled,
+      this.feedback,
+      this.childWhenDragging,
+      super.key,
+      required this.list});
 
   @override
   Widget build(BuildContext context) {
     return LongPressDraggable<(int, DraggableGridItem)>(
       onDraggableCanceled: (_, __) => onDragCancelled(),
       onDragStarted: () {
-        if (dragEnded) {
-          dragStarted = true;
-          dragEnded = false;
-          onDragEnded.call(false);
-          onDragStarted.call(true);
+        if (_dragEnded) {
+          _dragStarted = true;
+          _dragEnded = false;
         }
       },
       onDragEnd: (details) {
-        dragEnded = true;
-        dragStarted = false;
-        onDragEnded.call(true);
-        onDragStarted.call(false);
+        _dragEnded = true;
+        _dragStarted = false;
       },
       data: (index, list[index]),
       feedback: feedback ?? list[index].child,
       childWhenDragging:
-          childWhenDragging ?? draggedGridItem?.child ?? list[index].child,
+          childWhenDragging ?? _draggedGridItem?.child ?? list[index].child,
       child: list[index].child,
     );
   }